BASE-755 optimze bulk delete repositories

+++ src/main/kotlin/com/bitclave/node/repository/offer/OfferCrudRepository.kt
@@ -25,7 +25,7 @@ interface OfferCrudRepository : PagingAndSortingRepository<Offer, Long> {
 
     fun deleteByIdAndOwner(id: Long, owner: String): Long
 
-    fun deleteByOwner(owner: String): Long
+    fun deleteByOwner(owner: String): Int
 
     fun findByIdAndOwner(id: Long, owner: String): Offer?
